requirements.txt
requests~=2.32.4
PyYAML~=6.0.2
loguru~=0.7.3